Move parallel classes in their own package
